*, a: Add netlify-cli, setup netlify adapter
After switching to Gerrit, we can't use netlify to deploy from GitHub,
so we need to do it via the cli. This commit adds the cli to our flake
so it turns up in our dev shell.

It also adds the netlify adapter to 'a', which we need to locally build
the site. In future commits, we'll also do this for 'blog', as well as
the planned 'boba'

Change-Id: Ib7abf220f9877b263084220cb33df5b2bee73c2f
Signed-off-by: Skyler Grey <minion@clicks.codes>
Reviewed-on: https://git.clicks.codes/c/Minion/StarrySkySites/+/169