feat(a): Clean up top of page
Previously, we had a linux MOTD and python warranty disclaimer at
the top of the page, but this made it a bit more annoying to read to
the content and wasn't adding much. Also, the linux MOTD and python
disclaimer were a little bit weird together and I am not purely using
NixOS anymore.